Refactor: Move form-related components to ui/forms directory
- Moved form.tsx to ui/forms
- Moved label.tsx to ui/forms/label
- Moved input-otp.tsx to ui/forms/input
- Moved checkbox.tsx to ui/forms/checkbox
- Moved textarea.tsx to ui/forms/textarea
- Moved select.tsx to ui/forms/select
- Moved radio-group.tsx to ui/forms/radio
- Created comprehensive READMEs for each directory
- Updated original files with deprecation notices
- Updated COMPONENT-MOVEMENT-PLAN.md to track progress
