+++ src/components/ui/forms/checkbox/README.md
@@ -1,0 +1,13 @@
+# Checkbox Components
+
+This directory contains checkbox input components.
+
+## Components
+
+- `checkbox.tsx` - Basic checkbox component
+
+## Usage
+
+```tsx
+import { Checkbox } from '@/components/ui/forms/checkbox';
+```
