+++ src/components/ui/forms/checkbox/index.ts
@@ -1,0 +1,1 @@
+export { Checkbox } from './checkbox';
