+++ src/components/ui/forms/input/README.md
@@ -1,0 +1,15 @@
+# Input Components
+
+This directory contains components for text input fields.
+
+## Components
+
+- `base-input.tsx` - Basic text input component
+- `input-otp.tsx` - One-time password input component
+
+## Usage
+
+```tsx
+import { Input } from '@/components/ui/forms/input';
+import { InputOTP, InputOTPGroup, InputOTPSlot } from '@/components/ui/forms/input';
+```
