+++ src/components/ui/forms/input/index.ts
@@ -1,0 +1,2 @@
+export * from './input-otp';
+export * from './base-input';
