+++ src/components/ui/forms/label/index.ts
@@ -1,0 +1,1 @@
+export { Label } from './label';
