+++ src/components/ui/forms/radio/index.ts
@@ -1,0 +1,1 @@
+export { RadioGroup, RadioGroupItem } from './radio-group';
