+++ src/components/ui/forms/select/README.md
@@ -1,0 +1,18 @@
+# Select Components
+
+This directory contains dropdown/select input components.
+
+## Components
+
+- `select.tsx` - Select dropdown component with various subcomponents
+
+## Usage
+
+```tsx
+import { 
+  Select,
+  SelectTrigger,
+  SelectContent,
+  SelectItem 
+} from '@/components/ui/forms/select';
+```
