+++ src/components/ui/forms/textarea/README.md
@@ -1,0 +1,13 @@
+# Textarea Components
+
+This directory contains multiline text input components.
+
+## Components
+
+- `textarea.tsx` - Basic multiline text input component
+
+## Usage
+
+```tsx
+import { Textarea } from '@/components/ui/forms/textarea';
+```
