+++ src/components/ui/forms/textarea/index.ts
@@ -1,0 +1,1 @@
+export { Textarea } from './textarea';
